requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
faiss-cpu>=1.7.4
flask>=3.0.0
pydantic>=2.0.0
